Revert "Ditch insert_id in favor of Spanner-assigned timestamps for uniqueness."
This reverts commit bf8f07a642fd66ce333222e31dfec0ad0f45956b.

On second thought, it's likely that Spanner will be handling multiple batches of rollups from the same time period with duplicates inside the same transaction. Better to make the uniqueness explicit and depend on the known properties of PRNGs and the birthday paradox rather than be surprised by Beam, Dataflow, or Spanner.